Move Marshal/Unmarshal into ffjson package.
